@@ -1,5 +1,3 @@
-import 'package:flutter/foundation.dart';
-
 class Product {
   final String id;
   final String title;
@@ -9,11 +7,11 @@ class Product {
   bool isFavorite;
 
   Product({
-    @required this.id,
-    @required this.title,
-    @required this.description,
-    @required this.price,
-    @required this.imageUrl,
+    required this.id,
+    required this.title,
+    required this.description,
+    required this.price,
+    required this.imageUrl,
     this.isFavorite = false,
   });
 }
